src/subagent/system_designer_beta/requirements.txt
# Core dependencies
langgraph>=0.2.0
langchain>=0.3.0
langchain-anthropic>=0.3.0
anthropic>=0.40.0

# Configuration
python-dotenv>=1.0.0

# Data processing
pyyaml>=6.0
openpyxl>=3.1.0
python-docx>=1.1.0

# Logging
colorlog>=6.7.0

# Testing (optional)
pytest>=7.0.0
pytest-cov>=4.0.0
